requirements.txt
# Azure SDK and ML dependencies
azure-ai-ml>=1.12.0
azure-identity>=1.15.0
azure-core>=1.29.0

# Twitter API integration
tweepy>=4.14.0

# HTTP requests and utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Machine Learning and NLP
transformers>=4.35.0
torch>=2.1.0
datasets>=2.14.0
accelerate>=0.24.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Logging and monitoring
python-json-logger>=2.0.0

# Web framework (if building a web interface)
flask>=3.0.0
flask-cors>=4.0.0

# Optional: For advanced features
# openai>=1.3.0  # If using OpenAI models
# langchain>=0.0.350  # For advanced LLM orchestration
# chromadb>=0.4.0  # For vector database integration
